chore:extract common function
chore:extract common function
